refactor: update a disk offering list view
