+++ src/app/reducers/disk-offerings/redux/disk-offerings.reducers.ts
@@ -11,7 +11,6 @@ import * as event from './disk-offerings.actions';
 
 export interface State extends EntityState<DiskOffering> {
   loading: boolean;
-  tableParams: Array<string>;
 }
 
 export interface OfferingsState {
@@ -28,8 +27,7 @@ export const adapter: EntityAdapter<DiskOffering> = createEntityAdapter<DiskOffe
 });
 
 export const initialState: State = adapter.getInitialState({
-  loading: false,
-  tableParams: []
+  loading: false
 });
 
 export function reducer(
@@ -53,13 +51,6 @@ export function reducer(
       };
     }
 
-    case event.LOAD_DEFAULT_DISK_PARAMS_RESPONSE: {
-      return {
-        ...state,
-        tableParams: action.payload
-      }
-    }
-
     default: {
       return state;
     }
@@ -72,11 +63,6 @@ export const getOfferingsState = createFeatureSelector<OfferingsState>('disk-off
 export const getOfferingsEntitiesState = createSelector(
   getOfferingsState,
   state => state.list
-);
-
-export const getParams = createSelector(
-  getOfferingsEntitiesState,
-  state => state.tableParams
 );
 
 export const {
