+++ src/app/shared/actions/volume-actions/volume-resize/volume-resize.component.html
@@ -13,7 +13,6 @@
         <cs-disk-offering-selector
           name="diskOffering"
           [diskOfferings]="diskOfferings"
-          [params]="diskOfferingParams"
           [ngModel]="diskOffering"
           (change)="updateDiskOffering($event)"
         ></cs-disk-offering-selector>
