+++ src/app/shared/components/disk-offering/disk-offering-dialog/disk-offering-dialog.component.ts
@@ -1,8 +1,9 @@
 import { Component, Inject } from '@angular/core';
-import { MAT_DIALOG_DATA, MatDialogRef, MatTableDataSource } from '@angular/material';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material';
 import * as moment from 'moment';
 
 import { DiskOffering } from '../../../models';
+import { Utils } from '../../../services/utils/utils.service';
 
 @Component({
   selector: 'cs-disk-offering-dialog',
@@ -10,32 +11,32 @@ import { DiskOffering } from '../../../models';
   styleUrls: ['disk-offering-dialog.component.scss'],
 })
 export class DiskOfferingDialogComponent {
-  public diskOfferings: MatTableDataSource<DiskOffering>;
+  public diskOfferings: DiskOffering[];
   public selectedDiskOffering: DiskOffering;
-  public columns: Array<string>;
-  public columnsToDisplay: string[];
-  public fieldsToBeTranslated = ['provisioningtype', 'iscustomized'];
 
   constructor(
     @Inject(MAT_DIALOG_DATA) data,
     public dialogRef: MatDialogRef<DiskOfferingDialogComponent>,
   ) {
-    this.diskOfferings = new MatTableDataSource<DiskOffering>(data.diskOfferings);
+    this.diskOfferings = data.diskOfferings;
     this.selectedDiskOffering = data.diskOffering;
-    this.columns = data.columns;
-    this.columnsToDisplay = [...this.columns, 'radioButton'];
   }
 
   public offeringCreated(date: string): Date {
     return moment(date).toDate();
   }
 
-  public isElementInArray(column: string, columns: Array<string>): boolean {
-    return 0 <= columns.indexOf(column);
+  public convertToMb(bytes: number): number {
+    const megabytes = Utils.convertBytesToMegabytes(bytes);
+    return Math.round(megabytes);
   }
 
   public selectOffering(offering: DiskOffering) {
     this.selectedDiskOffering = offering;
+  }
+
+  public preventTriggerExpansionPanel(e: Event) {
+    e.stopPropagation(); // Don't open/close expansion panel when click on radio button
   }
 
   public onSubmit(): void {
@@ -44,7 +45,7 @@ export class DiskOfferingDialogComponent {
 
   public isSubmitButtonDisabled() {
     const isDiskOfferingNotSelected = !this.selectedDiskOffering;
-    const isNoDiskOfferings = !this.diskOfferings.data.length;
+    const isNoDiskOfferings = !this.diskOfferings.length;
     return isDiskOfferingNotSelected || isNoDiskOfferings;
   }
 }
