+++ src/app/shared/models/config/config.interface.ts
@@ -48,7 +48,6 @@ export interface CustomizableConfig {
   defaultComputeOffering: DefaultComputeOffering[];
   offeringCompatibilityPolicy: OfferingCompatibilityPolicy;
   serviceOfferingClasses: Array<any>;
-  diskOfferingParameters: Array<any>;
   offeringAvailability: OfferingAvailability;
 }
