+++ src/app/volume/container/volume-creation.container.ts
@@ -25,7 +25,6 @@ import { VolumeCreationData, VolumeType } from '../../shared/models/volume.model
       [diskOfferings]="offerings$ | async"
       [maxSize]="maxSize"
       [zones]="zones$ | async"
-      [params]="params$ | async"
       (onVolumeCreate)="createVolume($event)"
       (onZoneUpdated)="updateZone($event)"
     >
@@ -38,7 +37,6 @@ export class VolumeCreationContainerComponent extends WithUnsubscribe() implemen
   readonly offerings$ = this.store.select(fromDiskOfferings.selectAll);
   readonly zones$ = this.store.select(fromZones.selectAll);
   readonly account$ = this.store.select(fromAccounts.selectUserAccount);
-  readonly params$ = this.store.select(fromDiskOfferings.getParams);
 
   public maxSize = 2;
 
@@ -52,7 +50,6 @@ export class VolumeCreationContainerComponent extends WithUnsubscribe() implemen
 
   public ngOnInit() {
     this.store.dispatch(new diskOfferingActions.LoadOfferingsRequest({ type: VolumeType.DATADISK }));
-    this.store.dispatch(new diskOfferingActions.LoadDefaultParamsRequest());
   }
 
   public createVolume(data: VolumeCreationData) {
