+++ src/app/volume/volume-creation/volume-creation-dialog.component.html
@@ -41,7 +41,6 @@
         <cs-disk-offering-selector
           name="diskOffering"
           [diskOfferings]="diskOfferings"
-          [params]="params"
           [ngModel]="diskOffering"
           (change)="updateDiskOffering($event)"
         ></cs-disk-offering-selector>
