+++ src/app/volume/volume-creation/volume-creation-dialog.component.ts
@@ -1,9 +1,4 @@
-import {
-  Component,
-  EventEmitter,
-  Input,
-  Output
-} from '@angular/core';
+import { Component, EventEmitter, Input, Output } from '@angular/core';
 import { MatDialogRef } from '@angular/material';
 import { DiskOffering } from '../../shared/models/disk-offering.model';
 import { Zone } from '../../shared/models';
@@ -24,7 +19,6 @@ export class VolumeCreationDialogComponent {
   @Input() public diskOfferings: Array<DiskOffering>;
   @Input() public zones: Array<Zone>;
   @Input() public maxSize: number;
-  @Input() public params: Array<string>;
   @Output() public onVolumeCreate = new EventEmitter<VolumeCreationData>();
   @Output() public onZoneUpdated = new EventEmitter<Zone>();
