+++ src/config/config-example.json
@@ -142,16 +142,6 @@
       ]
     }
   ],
-  "diskOfferingParameters": [
-    "displaytext",
-    "disksize",
-    "created",
-    "storagetype",
-    "provisioningtype",
-    "iscustomized",
-    "miniops",
-    "maxiops"
-  ],
   "offeringAvailability": {
     "filterOfferings": true,
     "zones": {
