+++ src/i18n/en.json
@@ -617,6 +617,7 @@
       "fat": "Fat",
       "sparse": "Sparse",
       "shared": "Shared",
+      "local": "Local",
       "created": "Created",
       "true": "Yes",
       "false": "No"
@@ -1641,10 +1642,12 @@
     "GB": "GB",
     "GB_VALUE": "{{ value }} GB",
     "GHZ": "GHz",
-    "IOPS": "{{ value }} IO/s",
+    "IOPS": "IOPS",
+    "IOPS_VALUE": "{{ value }} IO/s",
     "MB": "MB",
     "MB_VALUE": "{{ value }} MB",
-    "MBPS": "{{ value }} MB/s",
+    "MBPS": "MB/s",
+    "MBPS_VALUE": "{{ value }} MB/s",
     "MHZ": "MHz"
   },
   "VOLUME_ACTIONS": {
@@ -1673,28 +1676,24 @@
   },
   "DISK_OFFERING": {
     "CHOOSE_DISK_OFFERING": "Choose disk offering",
-    "TABLE": {
-      "name": "Name",
-      "bytesreadrate": "Read MB/s",
-      "byteswriterate": "Write MB/s",
-      "iopsreadrate": "Read IO/s",
-      "iopswriterate": "Write IO/s",
-      "displaytext": "Description",
-      "disksize": "Disk size",
-      "storagetype": "Storage type",
-      "provisioningtype": "Provisioning type",
-      "iscustomized": "Customized",
-      "miniops": "Min IO/s",
-      "maxiops": "Max IO/s",
+    "BANDWIDTH_MBPS": "Bandwidth (MB/s)",
+    "READ": "Read",
+    "WRITE": "Write",
+    "MIN": "Min",
+    "MAX": "Max",
+    "STORAGE_TYPE": "Storage type",
+    "STORAGE_TYPE_VALUES": {
+      "shared": "Shared",
+      "local": "Local"
+    },
+    "PROVISIONING_TYPE": "Provisioning type",
+    "PROVISIONING_TYPE_VALUES": {
       "thin": "Thin",
       "thick": "Thick",
-      "fat": "Fat",
-      "sparse": "Sparse",
-      "shared": "Shared",
-      "created": "Created",
-      "true": "Yes",
-      "false": "No"
-    }
+      "fat": "Fat"
+    },
+    "CREATED": "Created",
+    "CUSTOM": "Custom"
   },
   "VM_PASSWORD": {
     "VM_PASSWORD": "Virtual machine password",
